requirements.txt
streamlit==1.28.1
supabase==2.1.0
python-dotenv==1.0.0
pandas==2.0.0
plotly==5.17.0
psycopg2-binary==2.9.9
streamlit-aggrid==0.3.4.post3